CI/CD with Github actions
When a PR request has finished, the push event on the dev branch
triggers an action which pushes the latest version of the branch to the
dev server. On the dev server, a post-recieve hook tears down the
existing docker containers, builds the latest version of the api, then
runs them again.

closes #284